backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-binance==1.0.19
pandas==2.2.2
ta==0.11.0
APScheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.0
